dlt-poc/mssql_pipeline/requirements.txt
dlt[athena,sql_database]>=1.6.0
aiobotocore>=2.15.0
s3fs>=2024.10.0
pyathena>=3.0.0
pymssql>=2.3.0
pyyaml>=6.0.1
boto3>=1.35.0
requests>=2.32.0
sqlalchemy>=2.0.0
tenacity>=8.0.0
